fix: align security hardening with latest main
